+++ spec/Permissions/PrivilegeSpec.php
@@ -1,6 +1,7 @@
 <?php
 namespace spec\BeatSwitch\Lock\Permissions;
 
+use BeatSwitch\Lock\Lock;
 use BeatSwitch\Lock\Resources\SimpleResource;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
@@ -18,16 +19,25 @@ class PrivilegeSpec extends ObjectBehavior
         $this->shouldImplement('BeatSwitch\Lock\Permissions\Permission');
     }
 
-    function it_can_validate_itself_against_parameters()
+    function it_can_validate_itself_against_parameters(Lock $lock)
     {
-        $this->isAllowed('edit', new SimpleResource('events', 1))->shouldReturn(true);
-        $this->isAllowed('edit', new SimpleResource('events', 1))->shouldReturn(true);
-        $this->isAllowed('edit', new SimpleResource('events', 2))->shouldReturn(false);
-        $this->isAllowed('delete', new SimpleResource('comments', 1))->shouldReturn(false);
+        $this->isAllowed($lock, 'edit', new SimpleResource('events', 1))->shouldReturn(true);
+        $this->isAllowed($lock, 'edit', new SimpleResource('events', 1))->shouldReturn(true);
+        $this->isAllowed($lock, 'edit', new SimpleResource('events', 2))->shouldReturn(false);
+        $this->isAllowed($lock, 'delete', new SimpleResource('comments', 1))->shouldReturn(false);
     }
 
     function it_can_match_an_equal_permission()
     {
         $this->matchesPermission($this)->shouldReturn(true);
+    }
+
+    function it_fails_with_a_false_condition(Lock $lock)
+    {
+        $this->beConstructedWith('edit', new SimpleResource('events', 1), function () {
+            return false;
+        });
+
+        $this->isAllowed($lock, 'edit', new SimpleResource('events', 1))->shouldReturn(false);
     }
 }