+++ src/Permissions/Condition.php
@@ -1,6 +1,7 @@
 <?php
 namespace BeatSwitch\Lock\Permissions;
 
+use BeatSwitch\Lock\Lock;
 use BeatSwitch\Lock\Resources\Resource;
 
 /**
@@ -12,9 +13,11 @@ interface Condition
     /**
      * Assert if the condition is correct
      *
+     * @param \BeatSwitch\Lock\Lock $lock
+     * @param \BeatSwitch\Lock\Permissions\Permission $permission
      * @param string $action
      * @param \BeatSwitch\Lock\Resources\Resource|null $resource
      * @return bool
      */
-    public function assert($action, Resource $resource = null);
+    public function assert(Lock $lock, Permission $permission, $action, Resource $resource = null);
 }