+++ src/Permissions/Restriction.php
@@ -1,6 +1,7 @@
 <?php
 namespace BeatSwitch\Lock\Permissions;
 
+use BeatSwitch\Lock\Lock;
 use BeatSwitch\Lock\Resources\Resource;
 
 /**
@@ -14,13 +15,14 @@ class Restriction extends AbstractPermission implements Permission
     /**
      * Validate a permission against the given params
      *
+     * @param \BeatSwitch\Lock\Lock $lock
      * @param string $action
      * @param \BeatSwitch\Lock\Resources\Resource|null $resource
      * @return bool
      */
-    public function isAllowed($action, Resource $resource = null)
+    public function isAllowed(Lock $lock, $action, Resource $resource = null)
     {
-        return ! $this->resolve($action, $resource);
+        return ! $this->resolve($lock, $action, $resource);
     }
 
     /**