+++ tests/Drivers/StaticDriverTestCase.php
@@ -16,8 +16,34 @@ abstract class StaticDriverTestCase extends PersistentDriverTestCase
     {
         $lock = $this->getCallerLock();
 
-        $lock->allow('upload', 'files', null, [new TrueConditionStub()]);
-        $lock->allow('upload', 'photos', null, [new FalseConditionStub()]);
+        $lock->allow('upload', 'files', null, new TrueConditionStub());
+        $lock->allow('upload', 'photos', null, new FalseConditionStub());
+
+        $this->assertTrue($lock->can('upload', 'files'));
+        $this->assertFalse($lock->can('upload', 'photos'));
+    }
+
+    /** @test */
+    final function it_can_work_with_multiple_conditions()
+    {
+        $lock = $this->getCallerLock();
+
+        $lock->allow('upload', 'files', null, [new FalseConditionStub(), new TrueConditionStub()]);
+
+        $this->assertFalse($lock->can('upload', 'files'));
+    }
+
+    /** @test */
+    final function it_can_work_with_callback_conditions()
+    {
+        $lock = $this->getCallerLock();
+
+        $lock->allow('upload', 'files', null, function () {
+            return true;
+        });
+        $lock->allow('upload', 'photos', null, function () {
+            return false;
+        });
 
         $this->assertTrue($lock->can('upload', 'files'));
         $this->assertFalse($lock->can('upload', 'photos'));